candidate/backend/requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
attrs==25.4.0
audioread==3.0.1
babel==2.17.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.3
coqpit-config==0.2.1
coqui-tts==0.27.2
coqui-tts-trainer==0.3.1
cycler==0.12.1
Cython==3.1.5
dateparser==1.1.8
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
fastapi==0.119.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
grpcio==1.76.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
inflect==7.5.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
lxml==6.0.2
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
networkx==3.5
num2words==0.5.14
numba==0.62.1
numpy==2.3.4
openai==2.6.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
PyPDF2==3.0.1
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
striprtf==0.0.29
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.55.4
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
